requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==4.0.3
attrs==25.1.0
av==14.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
dataclasses-json==0.6.7
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.17.0
filetype==1.2.0
flatbuffers==25.1.24
frozenlist==1.5.0
fsspec==2024.12.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.70.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.16
langchain-core==0.3.32
langchain-text-splitters==0.3.5
langsmith==0.3.2
livekit==0.19.1
livekit-agents==0.12.10
livekit-api==0.8.1
livekit-plugins-deepgram==0.6.16
livekit-plugins-elevenlabs==0.7.11
livekit-plugins-openai==0.10.16
livekit-plugins-silero==0.7.4
livekit-plugins-turn-detector==0.3.6
livekit-protocol==0.8.0
llama-index-core==0.12.14
llama-parse==0.5.20
lz4==4.4.3
MarkupSafe==3.0.2
marshmallow==3.26.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
onnxruntime==1.20.1
openai==1.60.2
orjson==3.10.15
packaging==24.2
pillow==11.1.0
pinecone==5.4.2
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
propcache==0.2.1
protobuf==4.25.6
protoc-gen-openapiv2==0.0.1
psutil==5.9.8
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
sympy==1.13.3
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.48.1
types-protobuf==4.25.0.20240417
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
watchfiles==0.24.0
wrapt==1.17.2
yarl==1.18.3
zstandard==0.23.0
